requirements.txt
python-telegram-bot
python-dotenv
firebase-admin
requests
gtts
httpx